Fix selection of language specific build type flags
In function llvmir_extract_lang_flags the selection of language flags had been
using the CMAKE_BUILD_TYPE variable. When CMAKE_BUILD_TYPE is dereferenced in a
debug or release build the variable is set to "Debug" or "Release". This commit
capitalizes the dereferenced content to make it compatible with the CMake
variable naming schema. For example: CMAKE_CXX_FLAGS_RELEASE or
CMAKE_CXX_FLAGS_DEBUG.